chore: updated login_page and moved it. renamed to landing_page, refactoring in progress
