+++ app/exerlog/lib/Bloc/authentication.dart
@@ -1,7 +1,7 @@
 import 'package:exerlog/Bloc/user_bloc.dart';
 import 'package:exerlog/Bloc/workout_bloc.dart';
 import 'package:exerlog/Models/user.dart';
-import 'package:exerlog/UI/login_screen/login_page.dart';
+import 'package:exerlog/src/feature/authentication/view/landing_screen.dart';
 import 'package:exerlog/UI/workout/workout_page.dart';
 import 'package:exerlog/main.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
